requirements.txt
APScheduler==3.10.4
babel==2.16.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-pubsub==2.23.1
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.66.1
idna==3.10
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rsa==4.9
six==1.16.0
tinydb==4.8.0
tzlocal==5.2
urllib3==2.2.3
